Moved jumbotron to front page only
