+++ src/components/Header/Header.jsx
@@ -3,17 +3,6 @@ import './Header.css';
 
 import TopBar from './TopBar';
 import SecondaryBar from './SecondaryBar';
-import images from '../../assets/Images';
-
-function Jumbotron() {
-	return (
-		<div className="jumbo">
-			<h2 className="title1">Masks as unique as you are.</h2>
-			<p></p>
-			<img src={images.boatMask.uri} alt={images.boatMask.altText}  />
-		</div>
-	)
-}
 
 class Header extends Component {
 	render() {
@@ -23,7 +12,6 @@ class Header extends Component {
 					placeholder='Search store'
 				/>
 				<SecondaryBar />
-				<Jumbotron />
 			</div>
 		)
 	}
